screg_pkg: Add package file for Space Cubics register design framework
This framework is intended to standardize and simplify the
implementation of register code in IP cores. By using this package,
designers can develop without needing to consider address assignments
or register field mappings.

Furthermore, by adhering to this framework, all register information
is centralized into a register table, enabling future enhancements
such as the addition of powerful debugging features based on this
data.

Signed-off-by: Daisuke Sasaki <daisuke@spacecubics.com>